Can now generate a palette mosaic image by giving a folder to be the image palette. Note that currently the palette images are not being scaled down, but are cropped.

+++ GroupGMosaicMaker/GroupGMosaicMaker/Model/TriangleGridGenerator.cs
@@ -14,7 +14,7 @@ namespace GroupGMosaicMaker.Model
         {
             base.DrawGrid(length);
             var height = 0;
-            for (var width = this.Decoder.PixelWidth - length; width < this.Decoder.PixelWidth; width += length)
+            for (var width = (int) this.Decoder.PixelWidth - length; width < this.Decoder.PixelWidth; width += length)
             {
                 height++;
                 this.SetPixelColor(width, height, Colors.White);
